+++ engine/cmodel_bsp.cpp
@@ -1199,7 +1199,7 @@ void CollisionBSPData_LoadDispInfo( CCollisionBSPData *pBSPData )
 	CDispTri  tempTris[MAX_DISPTRIS];
 
 	size_t nSize = 0;
-	int nCacheSize = 0;
+	size_t nCacheSize = 0;
 	int nPowerCount[3] = { 0, 0, 0 };
 
 	CMapLoadHelper lhDispInfo( LUMP_DISPINFO );
